testing CI with secrets

+++ NokiaHslVisualisation/app/src/main/java/com/xpyx/nokiahslvisualisation/data/TrafficItemRepository.kt
@@ -9,7 +9,7 @@ class TrafficItemRepository(private val trafficItemDao: TrafficItemDao) {
         trafficItemDao.addTrafficItem(trafficItem)
     }
 
-    suspend fun getTrafficItem(id: Int): TrafficItem {
+    suspend fun getTrafficItem(id: Long): TrafficItem {
         return trafficItemDao.getTrafficItem(id)
     }
